+++ .vscode/settings.json
@@ -17,7 +17,6 @@
     }
   },
   "ruff.format.preview": false,
-  "isort.check": true,
   "ruff.importStrategy": "useBundled",
   "ruff.organizeImports": false,
   "[javascript]": {
