+++ src/nonebot_plugin_parser_lite/data.py
@@ -7,9 +7,9 @@ from typing import Any, Literal, TypedDict
 
 from anyio import Path
 
-from ..download import DOWNLOADER
-from ..download.task import DownloadTaskWrapper
-from ..utils.ffmpeg import FFmpeg
+from .download import DOWNLOADER
+from .download.task import DownloadTaskWrapper
+from .utils.ffmpeg import FFmpeg
 
 
 def repr_path_task(path_task: DownloadTaskWrapper[Path]) -> str:
@@ -286,7 +286,7 @@ class ParseResult:
         for cont in self.content:
             if isinstance(cont, VideoContent):
                 return await cont.get_cover_path()
-            elif isinstance(cont, (ImageContent, GraphicContent)):
+            elif isinstance(cont, ImageContent | GraphicContent):
                 return await cont.get_path()
         return None
