+++ src/nonebot_plugin_parser_lite/parsers/__init__.py
@@ -1,9 +1,5 @@
-from .acfun import AcfunParser
-from .base import BaseParser, handle
-from .bilibili import BilibiliParser
-from .buff import BuffParser
-from .coolapk import CoolapkParser
-from .data import (
+from ..creator import Creator
+from ..data import (
     AudioContent,
     Author,
     GraphicContent,
@@ -12,6 +8,11 @@ from .data import (
     Platform,
     VideoContent,
 )
+from .acfun import AcfunParser
+from .base import BaseParser, handle
+from .bilibili import BilibiliParser
+from .buff import BuffParser
+from .coolapk import CoolapkParser
 from .douyin import DouyinParser
 from .duitang import DuiTangParser
 from .heybox import HeyBoxParser
@@ -37,6 +38,7 @@ __all__ = [
     "BilibiliParser",
     "BuffParser",
     "CoolapkParser",
+    "Creator",
     "DouyinParser",
     "DuiTangParser",
     "GraphicContent",
