+++ src/nonebot_plugin_parser_lite/parsers/illu/commentList.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 from datetime import datetime
 
 from msgspec import Struct, field
@@ -6,13 +8,13 @@ from msgspec.json import Decoder
 from .models import User
 
 
-class CommentObj(Struct):
+class Comment(Struct):
     author: User
     likeCount: int
     content: str
     objectId: str
     createdAt: str
-    subCommentList: list["CommentObj"] = field(default_factory=list)
+    subCommentList: list[Comment] = field(default_factory=list)
     subCommentCount: int = field(default=0)
 
     @property
@@ -23,7 +25,7 @@ class CommentObj(Struct):
 
 class CommentList(Struct):
     msg: str
-    results: list[CommentObj] = field(default_factory=list)
+    results: list[Comment] = field(default_factory=list)
 
 
 decoder = Decoder(CommentList)
