+++ src/nonebot_plugin_parser_lite/parsers/weibo/show_comment.py
@@ -4,8 +4,8 @@ import re
 from msgspec import Struct, field
 from msgspec.json import Decoder
 
+from ...creator import Creator
 from ...utils.format import replace_placeholder_to_sticker
-from ..creator import create_image, create_live_photo
 
 WEIBO_PATTERN = re.compile(r"\[(?P<name>[^]]+)\]")
 
@@ -25,12 +25,12 @@ class Pic(Struct):
     @property
     def content(self):
         if self.video:
-            return create_live_photo(
+            return Creator.live_photo(
                 video_url=self.video,
                 image_url=self.original.url,
                 ext_headers={"Referer": "https://weibo.com/"},
             )
-        return create_image(
+        return Creator.image(
             url=self.original.url, ext_headers={"Referer": "https://weibo.com/"}
         )
